PRSD-901: Extract builder methods
